requirements.txt
python-jose[cryptography]
passlib[bcrypt]
uvicorn
aiofiles
fastapi
python-multipart
bcrypt
pyblake2
pynacl
base58check
requests
pymongo # mongodb
dnspython # mongodb srv URI
python-dotenv
datetime
pytest